Dijkstra Labyrinth
This program, developed with Python, allows it to find the shortest path in the labyrinth. There are two different maze options in the program. The first one draws a map according to a matrix in the txt file. In this matrix, 0's are paths, 1, 2 and 3's are obstacles. The second is the option to generate a maze yourself. In this option, it forms a maze by branching towards the starting and ending notes. It finds the shortest path from the starting point to the target point using the Dijkstra algorithm in these two mazes.